requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
openai==1.51.2
httpx==0.27.2
langchain-core==0.2.41
langchain-openai==0.1.23
langgraph==0.2.15 
pinecone-client==5.0.0
langchain-pinecone==0.1.3
aiohttp==3.13.2

